+++ lib/l10n/app_localizations_de.dart
@@ -202,7 +202,7 @@ class AppLocalizationsDe extends AppLocalizations {
       'Fragen oder Fehlermeldungen? Öffne ein Issue im Projekt-Repository.';
 
   @override
-  String get gitSyncMarksAndroidIssues => 'GitSyncMarks-Mobile (GitHub Issues)';
+  String get gitSyncMarksAndroidIssues => 'GitSyncMarks-App (GitHub Issues)';
 
   @override
   String get profiles => 'Profile';