+++ lib/l10n/app_localizations_en.dart
@@ -199,7 +199,7 @@ class AppLocalizationsEn extends AppLocalizations {
       'Questions or error messages? Open an issue in the project repository.';
 
   @override
-  String get gitSyncMarksAndroidIssues => 'GitSyncMarks-Mobile (GitHub Issues)';
+  String get gitSyncMarksAndroidIssues => 'GitSyncMarks-App (GitHub Issues)';
 
   @override
   String get profiles => 'Profiles';