+++ lib/l10n/app_localizations_es.dart
@@ -201,7 +201,7 @@ class AppLocalizationsEs extends AppLocalizations {
       '¿Preguntas o mensajes de error? Abre un issue en el repositorio del proyecto.';
 
   @override
-  String get gitSyncMarksAndroidIssues => 'GitSyncMarks-Mobile (GitHub Issues)';
+  String get gitSyncMarksAndroidIssues => 'GitSyncMarks-App (GitHub Issues)';
 
   @override
   String get profiles => 'Perfiles';